requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
WTForms>=3.0.0
Werkzeug>=2.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pandas>=2.0.0
pymysql>=1.1.0
mysql-connector-python>=8.0.33

